services/auth_service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-cors-headers

pytest
pytest-django
pytest-cov
pytest-mock
